Reproducability Update
To make the game more machine playable, I removed all "realtime" aspects of the game, like total game time, and replaced them with steps. It's assumed that each FixedUpdate takes .02 seconds, and that 3000 steps will be simulated in a 60 seconds game. So, for the AI, the sequence of states will be the same for the same sequence of actions.

The only outstanding issue is that bullet collision seems to lag slightly when Time.timeScale is somewhere above 1. The result of this is that the game may progress for a few more steps, which is fine.