Feb 8 Checkpoint
Since last update, started using stable-baselines3 and specifically the SAC agent. I have 4 hand-coded opponents for SAC to train against, the stationary opponent, the random opponent, the line opponent (moves in a line up and down, shooting left), and the box agent (moves around the perimeter of the arena, shooting randomly). SAC agent is able to learn how to beat all of these agents >90% of the time.

I started working on a self-play training routine. It runs, but the learned behavior by the SAC agent is to win on the left and lose on the right. More work needs to be done with self-play.