Update 3.1 - PBT Bug Fixes and Tweaks
Fixed bug that incorrectly changed ELOs of agents
Changed running subprocesses with OS.system to subprocess.Popen
First attempt at training robustness. If a training process fails, it will be reran until it completes correctly.
Handling of NaN'ed models. If agents start to produce NaNs, training stops and the old model is saved as the newly trained model. So that agent will stop learning, but should eventually be killed off / replaced by the population.
Changed the game environment so that agents will start A) on either side randomly and B) at a random point on that side.