Update 0.3.6 PBT Hyperparameter Initialization and Bug Fix
Updated the preamble so that agent policies are initialized with hyperparameters that are close (but randomly slightly different) from stable-baselines3 PPO default values. Except for batch_size, which I default to 4096.

Also included a bug fix for training/eval code that closes and deletes the SubprocVecEnv object regardless of if the training/eval succeeds. I think this helps subsequent attempts at training/eval during the same iteration. Not 100% sure on this but it seems to work better

I also made the population stats plotting script better (plot_pop_elo.py)